+++ lib/europe/vat/vat.rb
@@ -4,75 +4,65 @@ require 'europe/vat/rates'
 require 'europe/vat/format'
 require 'uri'
 require 'net/http'
-require 'rexml/document'
+require 'json'
 require 'date'
 
 # Europe Gem
 module Europe
   # VAT
   module Vat
-    WSDL = 'http://ec.europa.eu/taxation_customs/vies/' \
-           'services/checkVatService'
+    API_URL = 'https://ec.europa.eu/taxation_customs/vies/rest-api/check-vat-number'
     HEADERS = {
-      'Content-Type' => 'text/xml;charset=UTF-8',
-      'SOAPAction' => ''
+      'Content-Type' => 'application/json',
+      'Accept' => 'application/json'
     }.freeze
 
-    BODY = <<-XML
-      <soapenv:Envelope
-      xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/"
-      xmlns:urn="urn:ec.europa.eu:taxud:vies:services:checkVat:types">
-        <soapenv:Header/>
-        <soapenv:Body>
-           <urn:checkVat>
-              <urn:countryCode>{COUNTRY_CODE}</urn:countryCode>
-              <urn:vatNumber>{NUMBER}</urn:vatNumber>
-           </urn:checkVat>
-        </soapenv:Body>
-      </soapenv:Envelope>
-    XML
+    VIES_ERRORS = {
+      'INVALID_INPUT' => :invalid_input,
+      'GLOBAL_MAX_CONCURRENT_REQ' => :rate_limited,
+      'MS_MAX_CONCURRENT_REQ' => :rate_limited,
+      'SERVICE_UNAVAILABLE' => :service_unavailable,
+      'MS_UNAVAILABLE' => :ms_unavailable,
+      'TIMEOUT' => :timeout
+    }.freeze
 
     def self.validate(number)
-      return :failed if number.size < 4
+      return :invalid_input if number.size < 4
 
       response = send_request(number[0..1], number[2..-1])
-      return :failed unless response.is_a? Net::HTTPSuccess
-      return :failed if response.body.include?('soap:Fault')
-      return :timeout if response.body.include?('TIMEOUT')
-      return :timeout if response.body.include?('MS_MAX_CONCURRENT_REQ')
+      return handle_error_response(response) unless response.is_a?(Net::HTTPSuccess)
 
       setup_response(response)
-    rescue Net::OpenTimeout
+    rescue Net::OpenTimeout, Net::ReadTimeout
       :timeout
-    rescue Net::HTTPServerError
-      :server_error
+    rescue Errno::ECONNREFUSED, Errno::ECONNRESET, Errno::ETIMEDOUT, SocketError, OpenSSL::SSL::SSLError
+      :service_unavailable
+    end
+
+    def self.handle_error_response(response)
+      body = JSON.parse(response.body)
+      error_code = body.dig('errorWrappers', 0, 'error')
+      VIES_ERRORS[error_code] || :service_error
+    rescue JSON::ParserError
+      :service_error
     end
 
     def self.setup_response(response)
-      body = response_xml(response)
+      body = JSON.parse(response.body)
       {
-        valid: extract_data(body, 3) == 'true',
-        country_code: extract_data(body, 0),
-        vat_number: extract_data(body, 1),
-        request_date: convert_date(extract_data(body, 2)),
-        name: extract_data(body, 4),
-        address: extract_data(body, 5)
+        valid: body['valid'] == true,
+        country_code: body['countryCode'],
+        vat_number: body['vatNumber'],
+        request_date: convert_date(body['requestDate']),
+        name: body['name'],
+        address: body['address']
       }
-    end
-
-    def self.response_xml(response)
-      xml = REXML::Document.new(response.body)
-      xml.elements.first.elements[2].elements[1]
     end
 
     def self.convert_date(date)
       return unless date
 
       Date.parse(date)
-    end
-
-    def self.extract_data(body, position)
-      body[position]&.text
     end
 
     def self.charge_vat?(origin_country, number)
@@ -83,18 +73,16 @@ module Europe
     end
 
     def self.send_request(country_code, number)
-      uri = URI.parse(WSDL)
+      uri = URI.parse(API_URL)
 
-      body = BODY.dup.gsub('{COUNTRY_CODE}', country_code)
-      body = body.gsub('{NUMBER}', number)
-
-      # Create the HTTP objects
-      http = Net::HTTP.new(uri.host, uri.port)
-      request = Net::HTTP::Post.new(uri.request_uri, HEADERS)
-      request.body = body
-
-      # Send the request
-      http.request(request)
+      Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
+        request = Net::HTTP::Post.new(uri.request_uri, HEADERS)
+        request.body = JSON.generate(
+          countryCode: country_code,
+          vatNumber: number
+        )
+        http.request(request)
+      end
     end
   end
 end